% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Using the greedy approach for the knapsack problem}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{Data frame with two variables v and w}

\item{W}{knapsack size}
}
\value{
maximum knapsack value and which elements
}
\description{
Using the greedy approach for the knapsack problem
}
