% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{Using the dynamic approach for the knapsack problem}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{Data frame with two variables v and w}

\item{W}{knapsack size}
}
\value{
maximum knapsack value and which elements
}
\description{
Using the dynamic approach for the knapsack problem
}
